+++ addOns/exim/src/main/java/org/zaproxy/addon/exim/automation/ImportJobDialog.java
@@ -42,13 +42,14 @@ public class ImportJobDialog extends StandardFieldsDialog {
     private static final String NAME_PARAM = "exim.automation.dialog.name";
     private static final String TYPE_PARAM = "exim.automation.dialog.type";
     private static final String FILE_NAME_PARAM = "exim.automation.dialog.filename";
+    private static final String SEND_REQUESTS_PARAM = "exim.automation.import.dialog.sendrequests";
 
     private ImportJob job;
 
     private DefaultComboBoxModel<ImportTypeOption> typeOptionModel;
 
     public ImportJobDialog(ImportJob job) {
-        super(View.getSingleton().getMainFrame(), TITLE, DisplayUtils.getScaledDimension(500, 200));
+        super(View.getSingleton().getMainFrame(), TITLE, DisplayUtils.getScaledDimension(500, 230));
         this.job = job;
 
         this.addTextField(NAME_PARAM, this.job.getData().getName());
@@ -72,6 +73,9 @@ public class ImportJobDialog extends StandardFieldsDialog {
             f = new File(fileName);
         }
         this.addFileSelectField(FILE_NAME_PARAM, f, JFileChooser.FILES_AND_DIRECTORIES, null);
+        this.addCheckBoxField(
+                SEND_REQUESTS_PARAM,
+                Boolean.TRUE.equals(this.job.getParameters().getSendRequests()));
         this.addPadding();
     }
 
@@ -101,6 +105,7 @@ public class ImportJobDialog extends StandardFieldsDialog {
         ImportTypeOption typeOption = (ImportTypeOption) typeOptionModel.getSelectedItem();
         this.job.getParameters().setType(typeOption.id().toLowerCase(Locale.ROOT));
         this.job.getParameters().setFileName(getStringValue(FILE_NAME_PARAM));
+        this.job.getParameters().setSendRequests(getBoolValue(SEND_REQUESTS_PARAM));
         this.job.resetAndSetChanged();
     }