+++ addOns/exim/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Added
+- Option to send requests when importing a HAR file, including via the Automation Framework and API.
 
 ## [0.21.0] - 2026-07-06
 ### Added